backend/requirements_v2.txt
astroid==2.14.1
async-timeout==4.0.2
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
dill==0.3.6
Flask==2.2.2
googleapis-common-protos==1.56.2
grpcio==1.51.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-aws-lambda==0.36b0
opentelemetry-instrumentation-dbapi==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-grpc==0.36b0
opentelemetry-instrumentation-jinja2==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-sqlite3==0.36b0
opentelemetry-instrumentation-urllib==0.36b0
opentelemetry-instrumentation-urllib3==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
platformdirs==3.0.0
protobuf==3.20.3
pylint==2.16.1
redis==4.5.1
requests==2.28.2
six==1.16.0
thrift==0.16.0
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
